% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rOstluft.data.R
\name{f}
\alias{f}
\title{returns path to file}
\usage{
f(fn)
}
\arguments{
\item{fn}{file name}
}
\value{
path to file
}
\description{
Wrapper around \code{\link[base:system.file]{base::system.file()}} to simplify generating file names
}
\examples{
path_to_file <- f("Zch_Stampfenbachstrasse_2010-2014.csv")

# is the same as
path_to_file <- system.file("extdata", "Zch_Stampfenbachstrasse_2010-2014.csv",
                            package = "rOstluft.data", mustWork = TRUE)
}
