% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rOstluft.data.R
\docType{package}
\name{rOstluft.data}
\alias{rOstluft.data}
\alias{rOstluft.data-package}
\title{rOstluft.data: Airquality data from zuerich for 2010-2014}
\description{
This Package contains air quality data from the measurement sites
Zuerich Stampfenbachstrasse, Zuerich Schimmelstrasse and Zuerich
Rosengartenstrasse for the years 2010 until 2014. The time resolution
is 30 minutes.
The goal for rOstluft.data is to provide data in a raw format for examples
and test of the package rOstluft.
}
\examples{
 path_to_file <- system.file("extdata", "Zch_Stampfenbachstrasse_2010-2014.csv",
    package = "rOstluft.data", mustWork = TRUE)
 print(path_to_file)

 list.files(system.file("extdata", package = "rOstluft.data"))

}
